Move section class to data_entry_flow
